+++ lib/src/element/sub/button/segmented_btn2_element.dart
@@ -78,6 +78,9 @@ class SegmentedBtn2Element extends MultiChildElement {
         params.containsKey(EnumSpWMLParams.allowEmpty)
             ? params[EnumSpWMLParams.allowEmpty]
             : SegmentedBtnParams.defEmptySelectionAllowed;
+    elParams.p.direction = params.containsKey(EnumSpWMLParams.axis)
+        ? params[EnumSpWMLParams.axis]
+        : Axis.horizontal;
     // SIDが設定されていなければエラー。
     if (getSID() == null) {
       throw SpWMLException(
@@ -259,6 +262,7 @@ class _SegmentedBtn2ElementWidgetState
       style: widget.elParams.p.style,
       showSelectedIcon: widget.elParams.p.showSelectedIcon,
       selectedIcon: widget.elParams.p.selectedIcon,
+      direction: widget.elParams.p.direction,
     );
   }
 }