+++ lib/src/element_params/sub/button/segmented_btn_params.dart
@@ -35,4 +35,5 @@ class SegmentedBtnParams {
   bool showSelectedIcon = true;
   Widget? selectedIcon;
   bool isEnabled = true;
+  Axis direction = Axis.horizontal;
 }